+++ backend/src/routes/authRoutes.ts
@@ -6,6 +6,7 @@ import {
   googleAuthRedirect,
   loginController,
   logoutController,
+  verifyEmail,
 } from '../controllers/authControllers';
 import { authenticate } from '../middlewares/authenticate';
 
@@ -17,5 +18,6 @@ authRouter.post('/signup/', createUser);
 authRouter.post('/login/', loginController);
 authRouter.get('/check-login/', authenticate, checkLogin);
 authRouter.post('/log-out/', authenticate, logoutController);
+authRouter.get('/verify-email', verifyEmail);
 
 export default authRouter;
