+++ backend/tsconfig.json
@@ -9,7 +9,7 @@
     "outDir": "./dist",
     "rootDir": "./src",
     "typeRoots": ["./src/@types", "./node_modules/@types"],
-    "types": ["node", "express"]
+    "types": ["node", "express"],
   },
   "include": ["/api/*.ts", "src/**/*.ts", "src/**/*.d.ts"],
   "exclude": ["node_modules"]
